+++ NEWS.md
@@ -1,5 +1,6 @@
 # fflr (development version)
 
+* Improve handling of pre-draft and no-history edge cases (#35).
 * Improve abbreviation techniques, add all roster slot and position IDs (#29).
 * Fix error checking in `ffl_api()`. Returns proper error message when there is
 a failure (e.g., non-public league) (#36).